Add null check to performTransformerOperation
Unlike the method getEntityFieldValue, performTransformerOperation does not contain a null check. This means that when the value in the database is null, the method will try to reverse transform null, which will throw an InvalidArgumentException. The solution is to add a null check to performTransformerOperation and simply return null when $oldValue is null (like in getEntityFieldValue).